requirements.txt
# djqango

Django==3.2
django-multiselectfield==0.1.12
djangorestframework==3.13.1
requests==2.28.0
bs4==0.0.1
beautifulsoup4==4.11.1